Refactor class names and package structure
The commit adjusts the program's structure by renaming "MergedAccessRuleService" to "AccessRuleService". It also relocates "OauthUserMatchingService" to a higher-level package and updates all references accordingly. This simplifies package structure and naming convention, improving code readability and maintenance.
