+++ pic-sure-auth-services/src/main/java/edu/harvard/hms/dbmi/avillach/auth/rest/AccessRuleController.java
@@ -2,7 +2,7 @@ package edu.harvard.hms.dbmi.avillach.auth.rest;
 
 import edu.harvard.hms.dbmi.avillach.auth.entity.AccessRule;
 import edu.harvard.hms.dbmi.avillach.auth.model.response.PICSUREResponse;
-import edu.harvard.hms.dbmi.avillach.auth.service.impl.authorization.MergedAccessRuleService;
+import edu.harvard.hms.dbmi.avillach.auth.service.impl.authorization.AccessRuleService;
 import io.swagger.v3.oas.annotations.*;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.annotation.security.RolesAllowed;
@@ -31,10 +31,10 @@ import static edu.harvard.hms.dbmi.avillach.auth.utils.AuthNaming.AuthRoleNaming
 @RequestMapping(value = "/accessRule")
 public class AccessRuleController {
 
-    private final MergedAccessRuleService accessRuleService;
+    private final AccessRuleService accessRuleService;
 
     @Autowired
-    public AccessRuleController(MergedAccessRuleService accessRuleService) {
+    public AccessRuleController(AccessRuleService accessRuleService) {
         this.accessRuleService = accessRuleService;
     }
