+++ pic-sure-auth-services/src/main/java/edu/harvard/hms/dbmi/avillach/auth/service/impl/authentication/OpenAuthenticationService.java
@@ -4,7 +4,7 @@ import edu.harvard.hms.dbmi.avillach.auth.entity.Role;
 import edu.harvard.hms.dbmi.avillach.auth.entity.User;
 import edu.harvard.hms.dbmi.avillach.auth.service.impl.RoleService;
 import edu.harvard.hms.dbmi.avillach.auth.service.impl.UserService;
-import edu.harvard.hms.dbmi.avillach.auth.service.impl.authorization.MergedAccessRuleService;
+import edu.harvard.hms.dbmi.avillach.auth.service.impl.authorization.AccessRuleService;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,10 +21,10 @@ public class OpenAuthenticationService {
 
     private final UserService userService;
     private final RoleService roleService;
-    private final MergedAccessRuleService accessRuleService;
+    private final AccessRuleService accessRuleService;
 
     @Autowired
-    public OpenAuthenticationService(UserService userService, RoleService roleService, MergedAccessRuleService accessRuleService) {
+    public OpenAuthenticationService(UserService userService, RoleService roleService, AccessRuleService accessRuleService) {
         this.userService = userService;
         this.roleService = roleService;
         this.accessRuleService = accessRuleService;
